split DatabaseRepository into two files
